// sdk/communication/communication-administration/src/phoneNumber/generated/src/phoneNumbersClient.ts
/*
 * Copyright (c) Microsoft Corporation.
 * Licensed under the MIT License.
 *
 * Code generated by Microsoft (R) AutoRest Code Generator.
 * Changes may cause incorrect behavior and will be lost if the code is regenerated.
 */

import * as operations from "./operations";
import * as Models from "./models";
import * as Mappers from "./models/mappers";
import { PhoneNumbersClientContext } from "./phoneNumbersClientContext";
import { PhoneNumbersClientOptionalParams } from "./models";

class PhoneNumbersClient extends PhoneNumbersClientContext {
  /**
   * Initializes a new instance of the PhoneNumbersClient class.
   * @param endpoint The endpoint of the Azure Communication resource.
   * @param options The parameter options
   */
  constructor(endpoint: string, options?: PhoneNumbersClientOptionalParams) {
    super(endpoint, options);
    this.phoneNumbers = new operations.PhoneNumbers(this);
  }

  phoneNumbers: operations.PhoneNumbers;
}

// Operation Specifications

export {
  PhoneNumbersClient,
  PhoneNumbersClientContext,
  Models as PhoneNumbersModels,
  Mappers as PhoneNumbersMappers
};
export * from "./operations";